Move RNG functions to their own header/source files
