+++ Source/drlg_l2.cpp
@@ -11,6 +11,7 @@
 #include "diablo.h"
 #include "drlg_l1.h"
 #include "engine/load_file.hpp"
+#include "engine/random.hpp"
 #include "objects.h"
 #include "player.h"
 #include "quests.h"
