+++ Source/items.cpp
@@ -17,6 +17,7 @@
 #include "dx.h"
 #include "engine/cel_sprite.hpp"
 #include "engine/load_cel.hpp"
+#include "engine/random.hpp"
 #include "engine/render/cel_render.hpp"
 #include "engine/render/text_render.hpp"
 #include "init.h"
