+++ Source/monster.cpp
@@ -18,6 +18,7 @@
 #include "drlg_l4.h"
 #include "engine/cel_header.hpp"
 #include "engine/load_file.hpp"
+#include "engine/random.hpp"
 #include "engine/render/cl2_render.hpp"
 #include "init.h"
 #include "lighting.h"
