+++ Source/multi.cpp
@@ -13,6 +13,7 @@
 #include "diablo.h"
 #include "dthread.h"
 #include "engine/point.hpp"
+#include "engine/random.hpp"
 #include "mainmenu.h"
 #include "nthread.h"
 #include "options.h"
