+++ Source/objects.cpp
@@ -13,6 +13,7 @@
 #include "drlg_l1.h"
 #include "drlg_l4.h"
 #include "engine/load_file.hpp"
+#include "engine/random.hpp"
 #include "error.h"
 #include "init.h"
 #include "lighting.h"
