+++ Source/player.cpp
@@ -11,6 +11,7 @@
 #include "dead.h"
 #include "engine/cel_header.hpp"
 #include "engine/load_file.hpp"
+#include "engine/random.hpp"
 #include "gamemenu.h"
 #include "init.h"
 #include "lighting.h"
