+++ Source/quests.cpp
@@ -10,6 +10,7 @@
 #include "control.h"
 #include "cursor.h"
 #include "engine/load_file.hpp"
+#include "engine/random.hpp"
 #include "engine/render/cel_render.hpp"
 #include "engine/render/text_render.hpp"
 #include "gendung.h"
