+++ Source/spells.cpp
@@ -8,6 +8,7 @@
 #include "control.h"
 #include "cursor.h"
 #include "engine/point.hpp"
+#include "engine/random.hpp"
 #include "gamemenu.h"
 #include "inv.h"
 #include "missiles.h"
