+++ Source/stores.cpp
@@ -11,6 +11,7 @@
 
 #include "cursor.h"
 #include "engine/load_cel.hpp"
+#include "engine/random.hpp"
 #include "engine/render/cel_render.hpp"
 #include "engine/render/text_render.hpp"
 #include "init.h"
