+++ Source/town.cpp
@@ -7,6 +7,7 @@
 
 #include "drlg_l1.h"
 #include "engine/load_file.hpp"
+#include "engine/random.hpp"
 #include "init.h"
 #include "player.h"
 #include "quests.h"
